feat(provider/kubernetes): v2 Add fetch manifest from cluster stage
While a recent change to the findArtifactsFromResource stage added
the manifest itself to the outputs, there are cases where a user may
not want to pollute the pipeline with an artifact, like the image
being used in the current deployment, but get only the manifest
itself.

This stage does that, and nothing more.